+++ front-main_page/src/pages/MUI/MediaCard.js
@@ -4,7 +4,7 @@ import CardActions from "@mui/material/CardActions";
 import CardContent from "@mui/material/CardContent";
 import CardMedia from "@mui/material/CardMedia";
 import Typography from "@mui/material/Typography";
-import baekjoon from "../image/boj-og.png";
+import baekjoon from "../../image/boj-og.png";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import Collapse from "@mui/material/Collapse";
 import IconButton from "@mui/material/IconButton";
